+++ src/ratings.h
@@ -29,11 +29,11 @@ class RatingsCalc
   absl::flat_hash_map<int, std::string_view> player_names_;
   std::vector<Player> player_info_;
   std::vector<Opponent> opponent_info_;
-  std::vector<size_t> game_indexes_;
-  std::vector<size_t> opp_index_;
-  std::vector<size_t> opp_played_;
+  std::vector<int> game_indexes_;
+  std::vector<int> opp_index_;
+  std::vector<int> opp_played_;
   std::vector<double> errors_;
-  std::vector<size_t> played_;
+  std::vector<int> played_;
   int next_player_ = 0;
   int games_ = 0;
 
@@ -48,7 +48,7 @@ class RatingsCalc
   void process_line(std::string_view line);
   double calculate_errors();
   void adjust_ratings_driver(int i, double e);
-  void calculate_errors(size_t start, size_t end);
+  void calculate_errors(int start, int end);
   void adjust_ratings(size_t start, size_t end);
   std::vector<ThreadPool::ThreadJob> create_error_calculation();
   std::vector<ThreadPool::ThreadJob> create_adjust_calculation();
